requirements.txt
appium-python-client
pytest
faker